refactor: consolidate memory resource
