+++ config/template.yaml
@@ -18,7 +18,7 @@ plugins:
       base_url: "${OLLAMA_BASE_URL}"
       model: "${OLLAMA_MODEL}"
     memory:
-      type: pipeline.resources.memory:Memory
+      type: entity.resources.memory:Memory
       dependencies: [database]
     storage:
       type: pipeline.resources.storage_resource:StorageResource
