+++ pyproject.toml
@@ -13,6 +13,7 @@ packages = [
     { include = "entity_config", from = "src" },
     { include = "registry", from = "src" },
     { include = "plugins", from = "src" },
+    { include = "entity", from = "src" },
     { include = "common_interfaces", from = "src" },
     { include = "tools", from = "src" },
     { include = "user_plugins" },
