+++ src/entity/__init__.py
@@ -1,0 +1,3 @@
+from .resources import Memory
+
+__all__ = ["Memory"]
