+++ src/entity/resources/__init__.py
@@ -1,0 +1,3 @@
+from .memory import Memory
+
+__all__ = ["Memory"]
