--- src/pipeline/resources/memory_filesystem.py
@@ -1,7 +1,0 @@
-from __future__ import annotations
-
-"""Wrapper for MemoryFileSystem."""
-
-from plugins.builtin.resources.memory_filesystem import MemoryFileSystem
-
-__all__ = ["MemoryFileSystem"]
