--- src/pipeline/resources/memory_storage.py
@@ -1,7 +1,0 @@
-from __future__ import annotations
-
-"""Wrapper for MemoryStorage."""
-
-from plugins.builtin.resources.memory_storage import MemoryStorage
-
-__all__ = ["MemoryStorage"]
