--- src/pipeline/resources/memory_vector_store.py
@@ -1,7 +1,0 @@
-from __future__ import annotations
-
-"""Wrapper for MemoryVectorStore."""
-
-from plugins.builtin.resources.memory_vector_store import MemoryVectorStore
-
-__all__ = ["MemoryVectorStore"]
