+++ src/plugins/builtin/resources/__init__.py
@@ -11,10 +11,8 @@ from .filesystem import FileSystemResource
 from .llm_base import LLM
 from .llm_resource import LLMResource
 from .local_filesystem import LocalFileSystemResource
-from .memory import Memory
+from entity.resources.memory import Memory
 from .memory_filesystem import MemoryFileSystem
-from .memory_storage import MemoryStorage
-from .memory_vector_store import MemoryVectorStore
 from .metrics import MetricsResource
 from .pg_vector_store import PgVectorStore
 from .s3_filesystem import S3FileSystem
@@ -36,8 +34,6 @@ __all__ = [
     "MetricsResource",
     "DuckDBDatabaseResource",
     "PgVectorStore",
-    "MemoryStorage",
-    "MemoryVectorStore",
     "MemoryFileSystem",
     "DuckDBVectorStore",
     "LocalFileSystemResource",
